+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "twitter-sentiment-analysis"
+name = "news-analyzer"
 version = "0.1.0"
 description = "Analyses tweets about a specified topic"
 readme = "README.md"
