+++ src/config.py
@@ -1,0 +1,2 @@
+TIME_PERIOD_MAP = {"Day": "d", "Week": "w", "Month": "m"}
+REGIONS_LANGUAGES = ["us-en", "ar-es", "es-es", "de-de"]
